fixed typo in painter

+++ src/ch/epfl/imhof/paintaing/Color.java
@@ -9,23 +9,23 @@ public final class Color {
     public static final Color WHITE = new Color(1, 1, 1);
     public static final Color BLACK = new Color(0, 0, 0);
 
-    private final float red;
-    private final float green;
-    private final float blue;
+    private final double red;
+    private final double green;
+    private final double blue;
 
-    private Color(float r, float g, float b) {
+    private Color(double r, double g, double b) {
         this.red = r;
         this.green = g;
         this.blue = b;
     }
 
-    public static Color gray(float level) {
+    public static Color gray(double level) {
         Preconditions.checkArgument(0 <= level && level <= 1);
 
         return new Color(level, level, level);
     }
 
-    public static Color rgb(float r, float g, float b) {
+    public static Color rgb(double r, double g, double b) {
         Preconditions.checkArgument(0 <= r && r <= 1);
         Preconditions.checkArgument(0 <= g && g <= 1);
         Preconditions.checkArgument(0 <= b && b <= 1);
@@ -34,9 +34,9 @@ public final class Color {
     }
 
     public static Color rgb(int values) {
-        float b = (values & 0xFF) / 255f;
-        float g = ((values >> 8) & 0xFF) / 255f;
-        float r = ((values >> 16) & 0xFF) / 255f;
+        double b = (values & 0xFF) / 255d;
+        double g = ((values >> 8) & 0xFF) / 255d;
+        double r = ((values >> 16) & 0xFF) / 255d;
 
         Preconditions.checkArgument(0 <= r && r <= 1);
         Preconditions.checkArgument(0 <= g && g <= 1);
@@ -45,15 +45,15 @@ public final class Color {
         return new Color(r, g, b);
     }
 
-    public float red() {
+    public double red() {
         return this.red;
     }
 
-    public float green() {
+    public double green() {
         return this.green;
     }
 
-    public float blue() {
+    public double blue() {
         return this.blue;
     }
 
@@ -64,7 +64,7 @@ public final class Color {
     }
 
     public static java.awt.Color convert(Color c) {
-        return new java.awt.Color(c.red(), c.green(), c.blue());
+        return new java.awt.Color((float) c.red(), (float) c.green(), (float) c.blue());
     }
 
 }
